+++ packages/asset-kit-block/src/helpers/assetKitHelpers.ts
@@ -48,3 +48,14 @@ export const blockStyle = (blockSetting: Settings): CSSProperties => {
 export const getSmallPreviewUrl = (url: string) => {
     return url.replace('{width}', '218');
 };
+
+export const transformDateStringToDate = (dateString: string) => {
+    const year = dateString.substring(0, 4);
+    const month = dateString.substring(4, 6);
+    const day = dateString.substring(6, 8);
+    const hour = dateString.substring(9, 11);
+    const minute = dateString.substring(11, 13);
+    const second = dateString.substring(13, 15);
+
+    return new Date(`${year}-${month}-${day}T${hour}:${minute}:${second}Z`);
+};
